+++ src/tedi/index.ts
@@ -55,3 +55,4 @@ export * from './providers/style-provider/style-provider';
 export * from './providers/accessibility-provider/accessibility-provider';
 export * from './providers/accessibility-provider/use-declare-loader';
 export * from './providers/theme-provider/theme-provider';
+export * from './providers/printing-provider/printing-provider';