+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# [17.0.0-rc.2](https://github.com/TEDI-Design-System/react/compare/react-17.0.0-rc.1...react-17.0.0-rc.2) (2026-04-14)
+
+
+### Bug Fixes
+
+* **variables:** update core version, update variable names [#592](https://github.com/TEDI-Design-System/react/issues/592) ([#598](https://github.com/TEDI-Design-System/react/issues/598)) ([1f15b36](https://github.com/TEDI-Design-System/react/commit/1f15b361dd3e953822c7fd490d772f4531363e0e))
+
+# [17.0.0-rc.1](https://github.com/TEDI-Design-System/react/compare/react-16.2.0-rc.1...react-17.0.0-rc.1) (2026-04-14)
+
+
+### Features
+
+* **print:** introduce PrintingProvider + context-based usePrint [#99](https://github.com/TEDI-Design-System/react/issues/99) ([#497](https://github.com/TEDI-Design-System/react/issues/497)) ([a58cb70](https://github.com/TEDI-Design-System/react/commit/a58cb70b49e715c9a24e37fe761d5b7ab4df3fa5))
+
+
+### BREAKING CHANGES
+
+* **print:** usePrint hook removed.
+Replace with usePrint from the new PrintingProvider context.
+
 # [16.2.0-rc.1](https://github.com/TEDI-Design-System/react/compare/react-16.1.1-rc.1...react-16.2.0-rc.1) (2026-04-06)